\name{EventList-class}
\docType{class}
\alias{EventList-class}
\alias{EventList}

\title{Class "EventList"}
\description{ Represents a Root Event List. This documentation is for
  internal use only. }
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("EventList", ...)}.
DO NOT USE THE new FUNCTION TO CREATE THIS OBJECT. Use
\code{\link{createEventList}} or \code{\link{makeEventList}} instead.
}
\section{Slots}{
	 \describe{
    \item{\code{ptr}:}{Object of class \code{"externalptr"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "EventList" in the signature.
}
\references{For more information on \emph{Root} see \url{http://root.cern.ch}.
Root is written by Rene Brun and Fons Rademakers. }
\author{ Adam L. Lyon }
\seealso{ \code{\link{createEventList}}, \code{\link{makeEventList}},
  \code{\link{addEventLists}} 
}
\examples{
}
\keyword{classes}
